qpid-bump,possibility to connect with failover and url-options (#577)
### Motivation
We use the connector as source function worker on OpenShift, to connect a queue on AMQBroker with a topic on Pulsar. Occasionally the pulsar sink pod loses connection, without crashing or sending any clear logs. 

### Modifications

To solve this issue we enable the connector to use qpids failover and keepalive option. This requires a uri that is prefixed and postfixed in a certain way. Furthermore the failover option allows for connection to more than one uri. 

To make this possible we have added a 'connection' configuration option which is backwards compatible:

using the connection configuration
- requires a list of uris (host, port, protocol and, optionally, urlOptions)
- a boolean flag to indicate whether failover is to be used.
- if connection is used, the current single host, protocol and port configuration cannot be used

It is possible to use 'connection', with useFailover set to false, a list with a single host, protocol, port and no url-options. This could replace the current single host, protocol and port configuration. That is why we have deprecated the current single host, protocol and port configuration

---------

Co-authored-by: Gerhard Bos <G.Bos@rechtspraak.nl>
Co-authored-by: Ronald Buit <ronald.buit@ndwcloud.nu>